saslserv: use size_t in all places taking a parameter size

+++ include/sasl.h
@@ -35,8 +35,8 @@ struct sasl_message_ {
 
 struct sasl_mechanism_ {
   char name[60];
-  int (*mech_start) (struct sasl_session_ *sptr, char **buffer, int *buflen);
-  int (*mech_step) (struct sasl_session_ *sptr, char *message, int length, char **buffer, int *buflen);
+  int (*mech_start) (struct sasl_session_ *sptr, char **buffer, size_t *buflen);
+  int (*mech_step) (struct sasl_session_ *sptr, char *message, size_t length, char **buffer, size_t *buflen);
   void (*mech_finish) (struct sasl_session_ *sptr);
 };
